Added a fair bit of flexibility to the reduction ops that can be performed on a FreqOutput list. This is useful particularly for OutputWithinRangeTC test comps.
Now, reduction ops can be written and used to accept an arbitrary number of
keyword arguments, taking advantage of Python's **kwargs functionality.

New reduction ops:
 * closestToSimTime
 * closestToStep
 * closestToVal (this can check closest to a value of both the target
    observable, or another observable).

Note: as part of this change, the ability to use the straight Python max()
and min() functions directly as reduction ops is no longer supported. However
I defined a maxOp and minOp in the stgfreq module, and this looks more natural anyway.

Also improved the unit tests for this functionality quite a bit.